+++ app/models/synchronization/adapter.rb
@@ -394,9 +394,9 @@ module CartoDB
         end
       rescue Exception => exception
         if exception.message.include?('canceling statement due to statement timeout')
-          # Check if the table has any lock
-          lock = user.in_database(as: :superuser).fetch(%Q{
-            SELECT pid, state, usename, query, query_start 
+          # Check if the table has any lock and cancel locking queries
+          locks = user.in_database(as: :superuser).fetch(%Q{
+            SELECT pid, query 
             FROM pg_stat_activity 
             WHERE pid in (
               SELECT pid FROM pg_locks l 
@@ -404,8 +404,14 @@ module CartoDB
               AND t.relkind = 'r' 
               WHERE t.relname IN ('#{table_name}')
             );
-          }).first
-          @logger.append_and_store "Transaction timed out as the table is blocked by query #{lock[:query]}. Retrying in 60 seconds..." if @logger && lock.present?
+          }).all
+          @logger.append_and_store "Transaction timed out as the table is blocked by other queries. Terminating locking queries and retrying in 60 seconds..." if @logger && locks.present?
+          locks.each do |lock|
+              @logger.append_and_store "Terminating query: #{lock[:query]}" if @logger
+              user.in_database(as: :superuser).execute %Q{
+              SELECT pg_terminate_backend(#{lock[:pid]});
+            }
+          end
           sleep(60) # wait 60 seconds and retry the swap
           database.transaction do
             rename(table_name, temporary_name) if exists?(table_name)